% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cucubes.R
\name{ComputeInterestingTuples}
\alias{ComputeInterestingTuples}
\title{Interesting tuples}
\usage{
ComputeInterestingTuples(acceleration.type = "scalar", dimensions = 1,
  divisions = 1, discretizations = 1, seed = 0, range = 1,
  pseudo.count = 0.001, reduce.method = "max", ig.thr,
  interesting.vars = c(), data, decision)
}
\arguments{
\item{acceleration.type}{acceleration type ('scalar' for none, 'avx'/'avx2' for use of the AVX/AVX2 instruction set respectively)}

\item{dimensions}{number of dimensions}

\item{divisions}{number of divisions}

\item{discretizations}{number of discretizations}

\item{seed}{seed for PRNG used during discretizations}

\item{range}{discretization range (from 0.0 to 1.0)}

\item{pseudo.count}{pseudo count}

\item{reduce.method}{discretization reduce method (either "max" or "mean")}

\item{ig.thr}{IG threshold above which the tuple is interesting}

\item{interesting.vars}{variables for which to check the IGs (none = all)}

\item{data}{input data where columns are variables and rows are observations}

\item{decision}{decision variable as a boolean vector of length equal to number of observations}
}
\value{
none (the function prints results)
}
\description{
Interesting tuples
}

