% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_relevant.R
\name{plot.MDFS}
\alias{plot.MDFS}
\title{Plot MDFS details}
\usage{
\method{plot}{MDFS}(x, plots = c("I", "p", "FDR"), ...)
}
\arguments{
\item{x}{an MDFS object}

\item{plots}{plots to plot (I for max IG, p for p-values, FDR, FWER)}

\item{...}{ignored}
}
\description{
Plot MDFS details
}

