% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_relevant.R
\name{RelevantVariables.MDFS}
\alias{RelevantVariables.MDFS}
\title{Find indices of relevant variables from MDFS}
\usage{
\method{RelevantVariables}{MDFS}(fs, level = 0.05, score = "FDR", ...)
}
\arguments{
\item{fs}{an MDFS object}

\item{level}{statistical significance level}

\item{score}{score to use}

\item{...}{ignored}
}
\value{
indices of relevant variables
}
\description{
Find indices of relevant variables from MDFS
}

